+++ lib/utils/constants/locale_constants.dart
@@ -1,4 +1,4 @@
-import '../common_libs.dart';
+import '../imports/common_libs.dart';
 
 class LocaleData {
   final Locale locale;
