+++ lib/view/component/ui/language_selector.dart
@@ -1,4 +1,4 @@
-import '../../../utils/common_libs.dart';
+import '../../../utils/imports/common_libs.dart';
 
 class LanguageSelector extends StatelessWidget {
   const LanguageSelector({super.key});
