+++ lib/view/component/ui/theme_toggle_button.dart
@@ -1,4 +1,4 @@
-import '../../../utils/common_libs.dart';
+import '../../../utils/imports/common_libs.dart';
 
 class ThemeToggleButton extends StatelessWidget {
   const ThemeToggleButton({super.key});
