+++ lib/view/theme/light_theme.dart
@@ -1,4 +1,4 @@
-import '../../../utils/common_libs.dart';
+import '../../utils/imports/common_libs.dart';
 
 ColorScheme _lightColorScheme = ColorScheme.light(
   // Primary colors
